worker/runner: Convert tests to integration tests
